scrap/requirements.txt
async-timeout==4.0.2
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
Deprecated==1.2.13
idna==3.3
kafka-python==2.0.2
numpy==1.21
packaging==21.3
pyparsing==3.0.7
redis==4.2.2
requests==2.27.1
soupsieve==2.3.1
urllib3==1.26.9
wrapt==1.14.0
apache-beam==2.37.0